+++ packages/subgraph/src/dao/utils.ts
@@ -1,9 +1,9 @@
 import {
-  Action,
   AddresslistVotingProposal,
   AdminProposal,
   MultisigProposal,
   TokenVotingProposal,
+  TransactionAction,
 } from '../../generated/schema';
 import {
   Executed,
@@ -23,6 +23,10 @@ import {handleERC20Action} from '../utils/tokens/erc20';
 import {handleERC721Action} from '../utils/tokens/erc721';
 import {handleERC1155Action} from '../utils/tokens/erc1155';
 import {handleNativeAction} from '../utils/tokens/eth';
+import {
+  generateTransactionActionEntityId,
+  generateDeterministicActionId,
+} from './ids';
 import {generateDaoEntityId} from '@aragon/osx-commons-subgraph';
 import {BigInt} from '@graphprotocol/graph-ts';
 
@@ -66,8 +70,13 @@ export function updateProposalWithFailureMap(
 export function handleAction<
   T extends ExecutedActionsStruct,
   R extends Executed
->(action: T, proposalId: string, index: i32, event: R): void {
-  let actionEntity = getOrCreateActionEntity(action, proposalId, index, event);
+>(action: T, transactionActionsId: string, index: i32, event: R): void {
+  let actionEntity = getOrCreateActionEntity(
+    action,
+    transactionActionsId,
+    index,
+    event
+  );
   actionEntity.execResult = event.params.execResults[index];
   actionEntity.save();
 
@@ -77,44 +86,63 @@ export function handleAction<
       action.to,
       action.value,
       'Native Token Withdraw',
-      proposalId,
+      transactionActionsId,
       index,
       event
     );
     return;
   }
 
-  handleTokenTransfers(action, proposalId, index, event);
+  handleTokenTransfers(action, transactionActionsId, index, event);
 }
 
 function getOrCreateActionEntity<
   T extends ExecutedActionsStruct,
   R extends Executed
->(action: T, proposalId: string, index: i32, event: R): Action {
-  const actionId = [proposalId, index.toString()].join('_');
-  let entity = Action.load(actionId);
+>(
+  action: T,
+  transactionActionsId: string,
+  index: i32,
+  event: R
+): TransactionAction {
+  const deterministicActionId = generateDeterministicActionId(
+    event.params.actor,
+    event.address,
+    event.params.callId,
+    index
+  );
+  const actionId = generateTransactionActionEntityId(
+    event.params.actor,
+    event.address,
+    event.params.callId,
+    index,
+    event.transaction.hash,
+    event.transactionLogIndex
+  );
 
-  // In case the execute on the dao is called by the address
-  // That we don't currently index for the actions in the subgraph,
-  // we fallback and still create an action.
-  // NOTE that it's important to generate action id differently to not allow
-
-  if (!entity) {
-    entity = new Action(actionId);
-    entity.to = action.to;
-    entity.value = action.value;
-    entity.data = action.data;
-    entity.proposal = proposalId;
-    entity.dao = generateDaoEntityId(event.address);
-  }
+  const entity = new TransactionAction(actionId);
+  entity.deterministicId = deterministicActionId;
+  entity.to = action.to;
+  entity.value = action.value;
+  entity.data = action.data;
+  entity.transactionActions = transactionActionsId;
+  entity.dao = generateDaoEntityId(event.address);
 
   return entity;
 }
 
+/**
+ * Determines if the action is an ERC20, ERC721 or ERC1155 transfer and calls the appropriate handler if so.
+ * Does nothing if the action is not a recognised token transfer.
+ * @param action the action to validate
+ * @param transactionActionsId the id container for a single set of executed actions
+ * @param actionIndex the index number of the action inside the executed batch
+ * @param event the Executed event emitting the event
+ */
 function handleTokenTransfers<
   T extends ExecutedActionsStruct,
   R extends Executed
->(action: T, proposalId: string, actionIndex: i32, event: R): void {
+>(action: T, transactionActionsId: string, actionIndex: i32, event: R): void {
   const methodSig = getMethodSignature(action.data);
 
   let handledByErc721: bool = false;
@@ -125,7 +153,7 @@ function handleTokenTransfers<
       action.to,
       event.address,
       action.data,
-      proposalId,
+      transactionActionsId,
       actionIndex,
       event
     );
@@ -136,7 +164,7 @@ function handleTokenTransfers<
       action.to,
       event.address,
       action.data,
-      proposalId,
+      transactionActionsId,
       actionIndex,
       event
     );
@@ -146,7 +174,7 @@ function handleTokenTransfers<
     handleERC20Action(
       action.to,
       event.address,
-      proposalId,
+      transactionActionsId,
       action.data,
       actionIndex,
       event