+++ packages/subgraph/tests/dao/dao_v1_3_0.test.ts
@@ -1,10 +1,16 @@
 import {
-  TransactionActionsProposal,
-  Action,
+  TransactionActions,
   ERC721Balance,
+  TransactionAction,
 } from '../../generated/schema';
 import {Executed} from '../../generated/templates/DaoTemplateV1_3_0/DAO';
 import {handleExecuted} from '../../src/dao/dao_v1_3_0';
+import {
+  generateTransactionActionEntityId,
+  generateDeterministicActionId,
+  generateTransactionActionsDeterministicId,
+  generateTransactionActionsEntityId,
+} from '../../src/dao/ids';
 import {GOVERNANCE_WRAPPED_ERC20_INTERFACE_ID} from '../../src/utils/constants';
 import {
   generateERC1155TransferEntityId,
@@ -32,6 +38,8 @@ import {
   TOKEN_SYMBOL,
   ERC20_TOTAL_SUPPLY,
   ERC20_DECIMALS,
+  ADDRESS_ONE,
+  ADDRESS_TWO,
 } from '../constants';
 import {
   ExtendedERC1155Balance,
@@ -46,12 +54,9 @@ import {
   getSupportsInterface,
 } from './utils';
 import {
-  generateActionEntityId,
   generateBalanceEntityId,
   generateDaoEntityId,
-  generateProposalEntityId,
   generateTokenIdBalanceEntityId,
-  generateTransactionActionsProposalEntityId,
   generateTransferEntityId,
   createDummyAction,
   createERC20TokenCalls,
@@ -80,7 +85,7 @@ describe('handleExecuted', () => {
     clearStore();
   });
 
-  test('successfuly creates action and proposal if not found', () => {
+  test('successfuly creates action and transactionAction', () => {
     let tuple: Array<ethereum.Value> = [ethereum.Value.fromString('')];
     let selector = '0x11111111';
 
@@ -103,56 +108,88 @@ describe('handleExecuted', () => {
 
     handleExecuted(event);
 
-    let proposalEntityId = generateProposalEntityId(
+    const deterministicID = generateTransactionActionsDeterministicId(
       event.params.actor,
-      BigInt.fromUnsignedBytes(event.params.callId)
+      event.address,
+      event.params.callId
     );
-    let transactionActionsProposalEntityId =
-      generateTransactionActionsProposalEntityId(
-        proposalEntityId,
-        event.transaction.hash,
-        event.transactionLogIndex
-      );
 
-    assert.entityCount('TransactionActionsProposal', 1);
-    assert.entityCount('Action', 2);
+    let transactionActionsEntityId = generateTransactionActionsEntityId(
+      event.params.actor,
+      event.address,
+      event.params.callId,
+      event.transaction.hash,
+      event.transactionLogIndex
+    );
+
+    assert.entityCount('TransactionActions', 1);
+    assert.entityCount('TransactionAction', 2);
 
     eq(
-      'TransactionActionsProposal',
-      transactionActionsProposalEntityId,
+      'TransactionActions',
+      transactionActionsEntityId,
       'id',
-      transactionActionsProposalEntityId
+      transactionActionsEntityId
     );
     eq(
-      'TransactionActionsProposal',
-      transactionActionsProposalEntityId,
+      'TransactionActions',
+      transactionActionsEntityId,
       'failureMap',
       failureMap
     );
     eq(
-      'TransactionActionsProposal',
-      transactionActionsProposalEntityId,
+      'TransactionActions',
+      transactionActionsEntityId,
       'allowFailureMap',
       allowFailureMap
     );
 
+    assert.fieldEquals(
+      'TransactionActions',
+      transactionActionsEntityId,
+      'deterministicId',
+      deterministicID
+    );
+
     for (let i = 0; i < event.params.actions.length; i++) {
-      let actionEntityId = generateActionEntityId(
-        transactionActionsProposalEntityId,
+      const deterministicActionID = generateDeterministicActionId(
+        event.params.actor,
+        event.address,
+        event.params.callId,
         i
       );
 
-      eq('Action', actionEntityId, 'id', actionEntityId);
-      eq('Action', actionEntityId, 'execResult', execResults[i].toHexString());
-      eq('Action', actionEntityId, 'dao', DAO_ADDRESS);
+      let actionEntityId = generateTransactionActionEntityId(
+        event.params.actor,
+        event.address,
+        event.params.callId,
+        i,
+        event.transaction.hash,
+        event.transactionLogIndex
+      );
+
+      eq('TransactionAction', actionEntityId, 'id', actionEntityId);
       eq(
-        'Action',
+        'TransactionAction',
         actionEntityId,
-        'proposal',
-        transactionActionsProposalEntityId
+        'execResult',
+        execResults[i].toHexString()
+      );
+      eq('TransactionAction', actionEntityId, 'dao', DAO_ADDRESS);
+      eq(
+        'TransactionAction',
+        actionEntityId,
+        'transactionActions',
+        transactionActionsEntityId
       );
       eq(
-        'Action',
+        'TransactionAction',
+        actionEntityId,
+        'deterministicId',
+        deterministicActionID
+      );
+      eq(
+        'TransactionAction',
         actionEntityId,
         'data',
         encodeWithFunctionSelector(tuple, selector).toHexString()
@@ -160,94 +197,52 @@ describe('handleExecuted', () => {
     }
   });
 
-  test('successfuly updates action and proposal if found', () => {
+  test('Duplicate deterministic ids will point to different proposals and actions', () => {
     let tuple: Array<ethereum.Value> = [ethereum.Value.fromString('')];
     let selector = '0x11111111';
-    let execResult = Bytes.fromHexString('0x11');
+
+    let execResults = [
+      Bytes.fromHexString('0x11'),
+      Bytes.fromHexString('0x22'),
+    ];
+
     let allowFailureMap = '2';
     let failureMap = '2';
 
-    let event = createExecutedEvent(
+    // event 0 has 2 actions
+    let event0 = createExecutedEvent(
+      [tuple, tuple],
+      [selector, selector],
+      allowFailureMap,
+      false,
+      execResults,
+      failureMap
+    );
+
+    // event1 has 1 action
+    let event1 = createExecutedEvent(
       [tuple],
       [selector],
       allowFailureMap,
       false,
-      [execResult],
+      execResults,
       failureMap
     );
 
-    let proposalEntityId = generateProposalEntityId(
-      event.params.actor,
-      BigInt.fromUnsignedBytes(event.params.callId)
-    );
-    let transactionActionsProposalEntityId =
-      generateTransactionActionsProposalEntityId(
-        proposalEntityId,
-        event.transaction.hash,
-        event.transactionLogIndex
-      );
+    // increment the log index to make sure the transaction actions are different
+    // for the two events even if they are in the same block
+    event1.logIndex = event1.logIndex.plus(BigInt.fromI32(1));
+    event1.transactionLogIndex = event1.logIndex;
 
-    let actionEntityId = generateActionEntityId(
-      transactionActionsProposalEntityId,
-      0
-    );
+    assert.entityCount('TransactionAction', 0);
+    assert.entityCount('TransactionActions', 0);
 
-    // create proposal
-    let proposal = new TransactionActionsProposal(
-      transactionActionsProposalEntityId
-    );
-    proposal.dao = event.address.toHexString();
-    proposal.createdAt = event.block.timestamp;
-    proposal.endDate = event.block.timestamp;
-    proposal.startDate = event.block.timestamp;
-    proposal.allowFailureMap = BigInt.fromString(allowFailureMap);
-    proposal.creator = event.params.actor;
-    proposal.executionTxHash = event.transaction.hash;
-    proposal.executed = true;
-    proposal.save();
-
-    // create action
-    let action = new Action(actionEntityId);
-    action.to = Address.fromString(DAO_TOKEN_ADDRESS);
-    action.data = Bytes.fromHexString('0x');
-    action.value = BigInt.zero();
-    action.dao = event.address.toHexString();
-    action.proposal = proposal.id;
-    action.save();
-
-    // Check that before `handleExecute`, execResults are empty
-    assert.entityCount('Action', 1);
-    assert.entityCount('TransactionActionsProposal', 1);
-    assert.assertTrue(action.execResult === null);
-    assert.assertTrue(proposal.failureMap === null);
-
-    handleExecuted(event);
+    handleExecuted(event0);
+    handleExecuted(event1);
 
     // The action and proposal count should be the same.
-    assert.entityCount('Action', 1);
-    assert.entityCount('TransactionActionsProposal', 1);
-
-    eq('Action', actionEntityId, 'id', actionEntityId);
-    eq('Action', actionEntityId, 'execResult', execResult.toHexString());
-
-    eq(
-      'TransactionActionsProposal',
-      transactionActionsProposalEntityId,
-      'id',
-      transactionActionsProposalEntityId
-    );
-    eq(
-      'TransactionActionsProposal',
-      transactionActionsProposalEntityId,
-      'failureMap',
-      failureMap
-    );
-    eq(
-      'TransactionActionsProposal',
-      transactionActionsProposalEntityId,
-      'allowFailureMap',
-      allowFailureMap
-    );
+    assert.entityCount('TransactionActions', 2);
+    assert.entityCount('TransactionAction', 3);
   });
 
   describe('ERC20 action', () => {
@@ -309,17 +304,13 @@ describe('handleExecuted', () => {
 
         handleExecuted(event);
 
-        let proposalEntityId = generateProposalEntityId(
+        let transactionActionsEntityId = generateTransactionActionsEntityId(
           event.params.actor,
-          BigInt.fromUnsignedBytes(event.params.callId)
+          event.address,
+          event.params.callId,
+          event.transaction.hash,
+          event.transactionLogIndex
         );
-        let transactionActionsProposalEntityId =
-          generateTransactionActionsProposalEntityId(
-            proposalEntityId,
-            event.transaction.hash,
-            event.transactionLogIndex
-          );
-
         let txHash = event.transaction.hash;
         let logIndex = event.transactionLogIndex;
         let timestamp = event.block.timestamp;
@@ -354,8 +345,8 @@ describe('handleExecuted', () => {
         eq(
           'ERC20Transfer',
           transferId,
-          'proposal',
-          transactionActionsProposalEntityId
+          'transactionActions',
+          transactionActionsEntityId
         );
         eq('ERC20Transfer', transferId, 'type', 'Withdraw');
         eq('ERC20Transfer', transferId, 'txHash', txHash.toHexString());
@@ -425,17 +416,13 @@ describe('handleExecuted', () => {
 
         handleExecuted(event);
 
-        let proposalEntityId = generateProposalEntityId(
+        let transactionActionsEntityId = generateTransactionActionsEntityId(
           event.params.actor,
-          BigInt.fromUnsignedBytes(event.params.callId)
+          event.address,
+          event.params.callId,
+          event.transaction.hash,
+          event.transactionLogIndex
         );
-        let transactionActionsProposalEntityId =
-          generateTransactionActionsProposalEntityId(
-            proposalEntityId,
-            event.transaction.hash,
-            event.transactionLogIndex
-          );
-
         let txHash = event.transaction.hash;
         let logIndex = event.transactionLogIndex;
         let timestamp = event.block.timestamp;
@@ -470,8 +457,8 @@ describe('handleExecuted', () => {
         eq(
           'ERC20Transfer',
           transferId,
-          'proposal',
-          transactionActionsProposalEntityId
+          'transactionActions',
+          transactionActionsEntityId
         );
         eq('ERC20Transfer', transferId, 'type', 'Withdraw');
         eq('ERC20Transfer', transferId, 'txHash', txHash.toHexString());
@@ -575,16 +562,13 @@ describe('handleExecuted', () => {
 
         handleExecuted(event);
 
-        let proposalEntityId = generateProposalEntityId(
+        let transactionActionsEntityId = generateTransactionActionsEntityId(
           event.params.actor,
-          BigInt.fromUnsignedBytes(event.params.callId)
+          event.address,
+          event.params.callId,
+          event.transaction.hash,
+          event.transactionLogIndex
         );
-        let transactionActionsProposalEntityId =
-          generateTransactionActionsProposalEntityId(
-            proposalEntityId,
-            event.transaction.hash,
-            event.transactionLogIndex
-          );
 
         // check ERC721Contract entity
         eq('ERC721Contract', tokenEntityId, 'id', tokenEntityId);
@@ -614,8 +598,8 @@ describe('handleExecuted', () => {
         eq(
           'ERC721Transfer',
           transferId,
-          'proposal',
-          transactionActionsProposalEntityId
+          'transactionActions',
+          transactionActionsEntityId
         );
         eq('ERC721Transfer', transferId, 'type', 'Withdraw');
         eq('ERC721Transfer', transferId, 'txHash', txHash.toHexString());
@@ -682,17 +666,13 @@ describe('handleExecuted', () => {
 
         handleExecuted(event);
 
-        let proposalEntityId = generateProposalEntityId(
+        let transactionActionsEntityId = generateTransactionActionsEntityId(
           event.params.actor,
-          BigInt.fromUnsignedBytes(event.params.callId)
+          event.address,
+          event.params.callId,
+          event.transaction.hash,
+          event.transactionLogIndex
         );
-        let transactionActionsProposalEntityId =
-          generateTransactionActionsProposalEntityId(
-            proposalEntityId,
-            event.transaction.hash,
-            event.transactionLogIndex
-          );
-
         // check ERC721Contract entity
         eq('ERC721Contract', tokenEntityId, 'id', tokenEntityId);
         eq('ERC721Contract', tokenEntityId, 'name', TOKEN_NAME);
@@ -721,8 +701,8 @@ describe('handleExecuted', () => {
         eq(
           'ERC721Transfer',
           transferId,
-          'proposal',
-          transactionActionsProposalEntityId
+          'transactionActions',
+          transactionActionsEntityId
         );
         eq('ERC721Transfer', transferId, 'type', 'Withdraw');
         eq('ERC721Transfer', transferId, 'txHash', txHash.toHexString());
@@ -839,23 +819,20 @@ describe('handleExecuted', () => {
           0,
           0
         );
-        let proposalEntityId = generateProposalEntityId(
+        let transactionActionsEntityId = generateTransactionActionsEntityId(
           event.params.actor,
-          BigInt.fromUnsignedBytes(event.params.callId)
+          event.address,
+          event.params.callId,
+          event.transaction.hash,
+          event.transactionLogIndex
         );
-        let transactionActionsProposalEntityId =
-          generateTransactionActionsProposalEntityId(
-            proposalEntityId,
-            event.transaction.hash,
-            event.transactionLogIndex
-          );
         assert.entityCount('ERC1155Transfer', 1);
         let erc1155Transfer = new ExtendedERC1155Transfer().withDefaultValues();
         erc1155Transfer.id = transferId;
         erc1155Transfer.amount = amount;
         erc1155Transfer.from = Address.fromHexString(daoEntityId);
         erc1155Transfer.to = Address.fromHexString(ADDRESS_THREE);
-        erc1155Transfer.proposal = transactionActionsProposalEntityId;
+        erc1155Transfer.transactionActions = transactionActionsEntityId;
         erc1155Transfer.type = 'Withdraw';
         erc1155Transfer.txHash = txHash;
         erc1155Transfer.createdAt = timestamp;
@@ -967,16 +944,13 @@ describe('handleExecuted', () => {
         // check ERC1155Transfer entity
         let txHash = event.transaction.hash;
         let logIndex = event.transactionLogIndex;
-        let proposalEntityId = generateProposalEntityId(
+        let transactionActionsEntityId = generateTransactionActionsEntityId(
           event.params.actor,
-          BigInt.fromUnsignedBytes(event.params.callId)
+          event.address,
+          event.params.callId,
+          event.transaction.hash,
+          event.transactionLogIndex
         );
-        let transactionActionsProposalEntityId =
-          generateTransactionActionsProposalEntityId(
-            proposalEntityId,
-            event.transaction.hash,
-            event.transactionLogIndex
-          );
         for (let i = 0; i < tokenIds.length; i++) {
           let erc1155Transfer =
             new ExtendedERC1155Transfer().withDefaultValues();
@@ -991,7 +965,7 @@ describe('handleExecuted', () => {
           erc1155Transfer.from = Address.fromHexString(daoEntityId);
           erc1155Transfer.to = Address.fromHexString(ADDRESS_THREE);
           erc1155Transfer.tokenId = tokenIds[i];
-          erc1155Transfer.proposal = transactionActionsProposalEntityId;
+          erc1155Transfer.transactionActions = transactionActionsEntityId;
           erc1155Transfer.type = 'Withdraw';
           erc1155Transfer.txHash = txHash;
           erc1155Transfer.createdAt = timestamp;
@@ -1049,6 +1023,64 @@ describe('handleExecuted', () => {
   });
 });
 
+describe('Testing ID generation', () => {
+  test('Deterministic ID generation', () => {
+    const caller = Address.fromString(ADDRESS_ONE);
+    const daoAddress = Address.fromString(ADDRESS_TWO);
+    const callId = '0xc0ffee';
+    const index = 255;
+
+    const actionId = generateDeterministicActionId(
+      caller,
+      daoAddress,
+      Bytes.fromHexString(callId),
+      index
+    );
+
+    assert.stringEquals(
+      actionId,
+      [
+        caller.toHexString(),
+        daoAddress.toHexString(),
+        callId,
+        index.toString(),
+      ].join('_')
+    );
+  });
+
+  test('ID generation with Tx data', () => {
+    const caller = Address.fromString(ADDRESS_ONE);
+    const daoAddress = Address.fromString(ADDRESS_TWO);
+    const callId = ZERO_BYTES32;
+    const index = 255;
+    const txHash = Bytes.fromHexString('0x1234567890abcdef');
+    const logIndex = BigInt.fromI32(12345);
+
+    const actionId = generateTransactionActionEntityId(
+      caller,
+      daoAddress,
+      Bytes.fromHexString(callId),
+      index,
+      txHash,
+      logIndex
+    );
+
+    // Assemblyscript can only concat if all are strings
+    // make sure your bytes are casted correctly
+    assert.stringEquals(
+      actionId,
+      [
+        caller.toHexString(),
+        daoAddress.toHexString(),
+        callId,
+        index.toString(),
+        txHash.toHexString(),
+        logIndex.toString(),
+      ].join('_')
+    );
+  });
+});
+
 // create Executed event with multiple actions
 function createExecutedEvent(
   tuple: ethereum.Value[][],
@@ -1072,7 +1104,7 @@ function createExecutedEvent(
       functionData.toHexString()
     );
 
-    actions.push(action);
+    actions.push(action as ethereum.Tuple);
   }
 
   if (execResults.length == 0) {